requirements.txt
alembic==1.13.3
bleach==6.1.0
blinker==1.8.2
Bootstrap-Flask==2.4.0
click==8.1.7
colorama==0.4.6
config==0.5.1
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
flask-bootstrap5==0.1.dev1
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.0.4
WTForms==3.1.2
